Add 'flexy-nowrap' row modifier.
When the auto columns, e.g. 'flexy-col-xs', go over the available space,
the columns are wrapped to the next line.  It might be more desirable
for the columns to instead shrink such that they all fit within the
available space.  This can now be done by adding the 'flexy-nowrap'
modifier to the same element as 'flexy-row'.